marked ol_x86_64 since it has a bug, and updated bmharness to prevent a 64-bit warning

+++ bench/bmharness.cpp
@@ -223,7 +223,7 @@ int main(int argc, char** argv) {
     pthread_attr_t attr;
     pthread_attr_init(&attr);
     pthread_attr_setscope(&attr, PTHREAD_SCOPE_SYSTEM);
-    for (uint32_t i = 0; i < CFG.threads; i++)
+    for (uintptr_t i = 0; i < CFG.threads; i++)
         args[i] = (void*)i;
 
     // actually create the threads
